# Parallelize git execution and set default value for RUST_LOG