--- a/Cargo.toml
+++ b/Cargo.toml
@@ -20,3 +20,4 @@ anyhow = "1.0.51"
 base64 = "0.13.0"
 pretty_env_logger = "0.4.0"
 clap = { version = "3.0.0-rc.7", features = ["derive", "env"] }
+futures = "0.3.19"