### Changes
* Made git updates truly asynchronous, by spawning green thread (Tokio tasks) and joining them
  * Added `futures` crate to easily join the tasks
* Set a default value for RUST_LOG (at first I thought there were no logging and found that strange 😅)

> The only thing that is not great with this change is the logging, since its now doing multiple repos at the same time, the logs are all over the places. Maybe this https://github.com/console-rs/indicatif could help

### Benchmark

The benchmark was run on the release binary and by simply using `time` followed by the binary and it was used on my account (which it cloned 39 repos).

|             | Initial run | Subsequent run |
| ------  | --------- | ---------------- |
| Before | 206.26s  | 47.915s               |
| After    | 34.48s    | 7.50s                   |